+++ src/components/Buttons/index.js
@@ -2,4 +2,4 @@ export { default as Button } from './Button'
 export { default as ArrowButton } from './ArrowButton'
 export { default as PublishButton } from './PublishButton'
 export { default as OrButton } from './OrButton'
-export { default as MoreButton } from './MoreButton'
+export { default as DropdownButton } from './DropdownButton'
